added tests and removed unused files, made small changes to some files
