+++ src/home/Home.css
@@ -1,5 +1,5 @@
 /* Container-Stil*/
-.con {
+.home-container {
   background-color: #fff;
 }
 
@@ -9,7 +9,6 @@
   width: 100%;
   height: 400px;
   display: flex;
-  background: url('saufisaufi.jpeg') top center/cover no-repeat;
   justify-content: center;
   align-items: center;
 }
